requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1
python-multipart==0.0.6
faster-whisper
sentence-transformers==2.2.2
faiss-cpu==1.8.0
bertopic==0.16.0
transformers==4.35.2
spacy==3.7.2
nltk==3.8.1
ffmpeg-python==0.2.0
yt-dlp
torch
